Unify data generator and the actual benchmark
