--- src/Akka.Persistence.Sql.Benchmark.Common/Akka.Persistence.Sql.Benchmark.Common.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>$(NetCoreTestVersion)</TargetFramework>
-        <Nullable>enable</Nullable>
-        <IsPackable>false</IsPackable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\Akka.Persistence.Sql.Tests.Common\Akka.Persistence.Sql.Tests.Common.csproj" />
-    </ItemGroup>
-
-</Project>
