+++ src/Akka.Persistence.Sql.Benchmarks/benchmark.conf
@@ -1,0 +1,5 @@
+
+benchmark {
+    connection-string = ""
+    provider-name = ""
+}
